remove error map function from PID main file
