+++ tests/test_pid.py
@@ -219,12 +219,22 @@ def test_converge_system():
 
 
 def test_error_map():
-    from simple_pid.PID import pi_clip
+   # error map function
+   def pi_clip(angle):
+      """Transform the angle value to a [-pi, pi) range."""
+      if angle > 0:
+         if angle > math.pi:
+            return angle - 2 * math.pi
+      else:
+         if angle < -math.pi:
+            return angle + 2 * math.pi
+      return angle
+
 
     sp = 0.  # setpoint
-    pid = PID(1, 0, 0, setpoint=sp, sample_time=0.1)
+    pid = PID(1, 0, 0, setpoint=sp, sample_time=0.1,
+       error_map=pi_clip) # include error mapping
     PV = 5.  # process variable
 
     # check if error value is mapped by the function
-    pid.error_map = pi_clip
     assert pid(PV) == pi_clip(sp - PV)  # clip the error
